+++ src/vue/modules/movements-history/store/index.spec.js
@@ -86,7 +86,6 @@ describe('movements-history.module', () => {
         const expectedParams = {
           page: {
             order: 'desc',
-            limit: 10,
           },
           filter: {
             account: 'GDIU5OQPAFPNBP75FQKMJTWSUKHTQTBTHXZWIZQR4DG4QRVJFPML6TTJ',
@@ -122,7 +121,6 @@ describe('movements-history.module', () => {
         const expectedParams = {
           page: {
             order: 'desc',
-            limit: 10,
           },
           filter: {
             asset: 'TTK',