+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 ## [Unreleased]
+#### Added
+- Get data from store for general kyc form
+- A loader on kyc recovery management page
+
 #### Fixed
 - Localization mistakes
 - A bug with trailing digits count on amount fields
+- Missing Russian localization
+- A bug with not show more button on tx history
 
 ### "Under the hood" changes
 #### Changed